safari-600.7-branch/LayoutTests/editing/spelling/spelling-unified-emulation-expected.txt

The spellchecker should work correctly if unified spellcheck path is used.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS numMarked is length
PASS str is markedStr
PASS numMarked is length
PASS str is markedStr
PASS numMarked is length
PASS str is markedStr
PASS str is markedStr
PASS numMarked is length
PASS str is markedStr
PASS str is markedStr
PASS str is markedStr
PASS numMarked is length
PASS str is markedStr
PASS str is markedStr
PASS str is markedStr
PASS numMarked is length
PASS str is markedStr
PASS numMarked is length
PASS str is markedStr
PASS numMarked is length
PASS str is markedStr
PASS numMarked is length
PASS numMarked is length
PASS successfullyParsed is true

TEST COMPLETE
I would like to sleep, buzz.
